FEAT: box styling

+++ client/src/Components/UserSettings/UserSettings.js
@@ -1,10 +1,12 @@
 import React, { useState } from "react";
 import axios from "axios";
+import styles from './UserSettings.module.css'
 
 function UserSettings(props) {
   const user = JSON.parse(localStorage.getItem("user"));
   const [selectedFile, setSelectedFile] = useState(null);
   const [changed, setChanged] = useState(false);
+  const [error, setError] = useState('')
 
   const fileSelectedHandler = (e) => {
     e.preventDefault();
@@ -12,8 +14,13 @@ function UserSettings(props) {
   };
 
   const fileUploadHandler = () => {
+    if(!selectedFile) {
+      setError("Vennligst velg et bilde først")
+      return
+    }
     const fd = new FormData();
     fd.append("image", selectedFile, selectedFile.name);
+    console.log(user.image)
     axios
       .put("/users/" + user.id + "/image", fd, {
         onDownloadProgress: (progressEvent) => {
@@ -26,8 +33,8 @@ function UserSettings(props) {
       })
       .then((res) => {
         // Delete the old pic from /images to save space
-        const imageName = user.image;
-        axios.delete("/images/" + imageName).catch((err) => console.log(err));
+        const imageName = JSON.parse(localStorage.getItem("user")).image;
+        if(imageName) axios.delete("/images/" + imageName).catch((err) => console.log(err));
 
         // Set the new image in localStorage
         const user = JSON.parse(localStorage.getItem("user"));
@@ -43,13 +50,18 @@ function UserSettings(props) {
     <div>
       <h1>Innstillinger</h1>
       <p>Velg et nytt profilbilde</p>
-      <input type="file" onChange={fileSelectedHandler} />
-      <br />
-      <br />
+      <label className={"file-input " + styles.FileInput} for="image-input">
+          <i class="fas fa-upload"></i> Trykk for å velge bilde
+        </label>
+        {selectedFile && <span id="file-selected">{selectedFile.name}</span>}
+      <input id="image-input" type="file" onChange={fileSelectedHandler} />
+      <br/>
+      <br/>
       <button className="button" onClick={fileUploadHandler}>
         Last opp
       </button>
       {changed && <p>Profilbilde endret!</p>}
+      {error && <p>{error}</p>}
       <br/>
       <br/>
     </div>
